+++ playwright/integrations/authenticated-chat-with-transcripts.spec.ts
@@ -1,69 +1,43 @@
 import fetchOmnichannelConfig from '../utils/fetchOmnichannelConfig';
 import fetchTestPageUrl from '../utils/fetchTestPageUrl';
+import fetchAuthUrl from '../utils/fetchAuthUrl';
 import { test, expect } from '@playwright/test';
 import OmnichannelEndpoints from '../utils/OmnichannelEndpoints';
 
 const testPage = fetchTestPageUrl();
-const omnichannelConfig = fetchOmnichannelConfig('UnauthenticatedChatWithTranscripts');
+const omnichannelConfig = fetchOmnichannelConfig('AuthenticatedChatWithTranscripts');
+const authUrl = fetchAuthUrl('AuthenticatedChatWithTranscripts');
 
-test.describe('@UnauthenticatedChat @UnauthenticatedChatWithTranscripts', () => {
-    test('ChatSDK.emailLiveChatTranscript() should not fail', async ({ page }) => {
-        await page.goto(testPage);
-
-        const [request, response, runtimeContext] = await Promise.all([
-            page.waitForRequest(request => {
-                return request.url().includes(OmnichannelEndpoints.LiveChatTranscriptEmailRequestPath);
-            }),
-            page.waitForResponse(response => {
-                return response.url().includes(OmnichannelEndpoints.LiveChatTranscriptEmailRequestPath);
-            }),
-            await page.evaluate(async ({ omnichannelConfig }) => {
-                const {OmnichannelChatSDK_1: OmnichannelChatSDK} = window;
-                const chatSDK = new OmnichannelChatSDK.default(omnichannelConfig);
-                const runtimeContext = {};
-
-                await chatSDK.initialize();
-
-                await chatSDK.startChat();
-
-                runtimeContext.requestId = chatSDK.requestId;
-
-                const body = {
-                    emailAddress: 'contoso@microsoft.com',
-                    attachmentMessage: 'Attachment Message'
-                };
-
-                await chatSDK.emailLiveChatTranscript(body);
-
-                await chatSDK.endChat();
-
-                return runtimeContext;
-            }, { omnichannelConfig })
-        ]);
-
-        const {requestId} = runtimeContext;
-        const requestUrl = `${omnichannelConfig.orgUrl}/${OmnichannelEndpoints.LiveChatTranscriptEmailRequestPath}/${requestId}?channelId=lcw`;
-
-        expect(request.url() === requestUrl).toBe(true);
-        expect(response.status()).toBe(200);
-    });
-
+test.describe('@AuthenticatedChat @AuthenticatedChatWithTranscripts', () => {
     test('ChatSDK.getLiveChatTranscript() should not fail', async ({ page }) => {
         await page.goto(testPage);
 
         const [request, response, runtimeContext] = await Promise.all([
             page.waitForRequest(request => {
-                return request.url().includes(OmnichannelEndpoints.LiveChatv2GetChatTranscriptPath);
+                return request.url().includes(OmnichannelEndpoints.LiveChatv2AuthGetChatTranscriptPath);
             }),
             page.waitForResponse(response => {
-                return response.url().includes(OmnichannelEndpoints.LiveChatv2GetChatTranscriptPath);
+                return response.url().includes(OmnichannelEndpoints.LiveChatv2AuthGetChatTranscriptPath);
             }),
-            await page.evaluate(async ({ omnichannelConfig }) => {
+            await page.evaluate(async ({ omnichannelConfig, authUrl }) => {
                 const {OmnichannelChatSDK_1: OmnichannelChatSDK} = window;
-                const chatSDK = new OmnichannelChatSDK.default(omnichannelConfig);
+
+                const payload = {
+                    method: "POST"
+                };
+
+                const response = await fetch(authUrl, payload);
+                const authToken = await response.text();
+
+                const chatSDKConfig = {
+                    getAuthToken: () => authToken
+                };
+
+                const chatSDK = new OmnichannelChatSDK.default(omnichannelConfig, chatSDKConfig);
                 const runtimeContext = {};
 
                 await chatSDK.initialize();
+                runtimeContext.authToken = authToken;
 
                 await chatSDK.startChat();
                 runtimeContext.requestId = chatSDK.requestId;
@@ -77,17 +51,18 @@ test.describe('@UnauthenticatedChat @UnauthenticatedChatWithTranscripts', () => 
                 await chatSDK.endChat();
 
                 return runtimeContext;
-            }, { omnichannelConfig })
+            }, { omnichannelConfig, authUrl })
         ]);
 
-        const {requestId, token, chatId, transcript} = runtimeContext;
-        const requestUrl = `${omnichannelConfig.orgUrl}/${OmnichannelEndpoints.LiveChatv2GetChatTranscriptPath}/${chatId}/${requestId}?channelId=lcw`;
+        const {authToken, requestId, token, chatId, transcript} = runtimeContext;
+        const requestUrl = `${omnichannelConfig.orgUrl}/${OmnichannelEndpoints.LiveChatv2AuthGetChatTranscriptPath}/${chatId}/${requestId}?channelId=lcw`;
         const requestHeaders = request.headers();
 
         expect(request.url() === requestUrl).toBe(true);
         expect(requestHeaders['authorization']).toBe(token);
         expect(requestHeaders['organizationid']).toBe(omnichannelConfig.orgId);
         expect(requestHeaders['widgetappid']).toBe(omnichannelConfig.widgetId);
+        expect(requestHeaders['authenticatedusertoken']).toBe(authToken);
         expect(response.status()).toBe(200);
         expect(Object.keys(transcript).includes('chatMessagesJson'));
         expect(typeof transcript['chatMessagesJson']).toBe('string');
@@ -104,17 +79,30 @@ test.describe('@UnauthenticatedChat @UnauthenticatedChatWithTranscripts', () => 
 
         const [request, response, runtimeContext] = await Promise.all([
             page.waitForRequest(request => {
-                return request.url().includes(OmnichannelEndpoints.LiveChatv2GetChatTranscriptPath);
+                return request.url().includes(OmnichannelEndpoints.LiveChatv2AuthGetChatTranscriptPath);
             }),
             page.waitForResponse(response => {
-                return response.url().includes(OmnichannelEndpoints.LiveChatv2GetChatTranscriptPath);
+                return response.url().includes(OmnichannelEndpoints.LiveChatv2AuthGetChatTranscriptPath);
             }),
-            await page.evaluate(async ({ omnichannelConfig }) => {
+            await page.evaluate(async ({ omnichannelConfig, authUrl }) => {
                 const {OmnichannelChatSDK_1: OmnichannelChatSDK} = window;
-                const chatSDK = new OmnichannelChatSDK.default(omnichannelConfig);
+
+                const payload = {
+                    method: "POST"
+                };
+
+                const response = await fetch(authUrl, payload);
+                const authToken = await response.text();
+
+                const chatSDKConfig = {
+                    getAuthToken: () => authToken
+                };
+
+                const chatSDK = new OmnichannelChatSDK.default(omnichannelConfig, chatSDKConfig);
                 const runtimeContext = {};
 
                 await chatSDK.initialize();
+                runtimeContext.authToken = authToken;
 
                 await chatSDK.startChat();
 
@@ -130,17 +118,18 @@ test.describe('@UnauthenticatedChat @UnauthenticatedChatWithTranscripts', () => 
                 runtimeContext.transcript = transcript;
 
                 return runtimeContext;
-            }, { omnichannelConfig })
+            }, { omnichannelConfig, authUrl })
         ]);
 
-        const {requestId, token, chatId, transcript} = runtimeContext;
-        const requestUrl = `${omnichannelConfig.orgUrl}/${OmnichannelEndpoints.LiveChatv2GetChatTranscriptPath}/${chatId}/${requestId}?channelId=lcw`;
+        const {authToken, requestId, token, chatId, transcript} = runtimeContext;
+        const requestUrl = `${omnichannelConfig.orgUrl}/${OmnichannelEndpoints.LiveChatv2AuthGetChatTranscriptPath}/${chatId}/${requestId}?channelId=lcw`;
         const requestHeaders = request.headers();
 
         expect(request.url() === requestUrl).toBe(true);
         expect(requestHeaders['authorization']).toBe(token);
         expect(requestHeaders['organizationid']).toBe(omnichannelConfig.orgId);
         expect(requestHeaders['widgetappid']).toBe(omnichannelConfig.widgetId);
+        expect(requestHeaders['authenticatedusertoken']).toBe(authToken);
         expect(response.status()).toBe(200);
         expect(Object.keys(transcript).includes('chatMessagesJson'));
         expect(typeof transcript['chatMessagesJson']).toBe('string');