+++ playwright/integrations/unauthenticated-chat.spec.ts
@@ -22,7 +22,6 @@ test.describe('UnauthenticatedChat @UnauthenticatedChat', () => {
                 const { OmnichannelChatSDK_1: OmnichannelChatSDK } = window;
                 const chatSDK = new OmnichannelChatSDK.default(omnichannelConfig);
 
-                chatSDK.setDebug(true);
                 await chatSDK.initialize();
 
                 const runtimeContext = {
@@ -650,5 +649,111 @@ test.describe('UnauthenticatedChat @UnauthenticatedChat', () => {
         expect(sessionInitRequestPostData.device).not.toBe(null);
         expect(sessionInitRequestPostData.os).not.toBe(null);
         expect(sessionInitRequestPostData.originurl).toBe(testPage);
+    });
+
+    test('ChatSDK.getConversationDetails() should not fail', async ({page}) => {
+        await page.goto(testPage);
+
+        const [liveWorkItemDetailsRequest, liveWorkItemDetailsResponse, runtimeContext] = await Promise.all([
+            page.waitForRequest(request => {
+                return request.url().includes(OmnichannelEndpoints.LiveChatLiveWorkItemDetailsPath);
+            }),
+            page.waitForResponse(response => {
+                return response.url().includes(OmnichannelEndpoints.LiveChatLiveWorkItemDetailsPath);
+            }),
+            await page.evaluate(async ({ omnichannelConfig }) => {
+                const { OmnichannelChatSDK_1: OmnichannelChatSDK } = window;
+                const chatSDK = new OmnichannelChatSDK.default(omnichannelConfig);
+
+                await chatSDK.initialize();
+
+                await chatSDK.startChat();
+
+                const conversationDetails = await chatSDK.getConversationDetails();
+
+                const runtimeContext = {
+                    requestId: chatSDK.requestId,
+                    conversationDetails
+                };
+
+                await chatSDK.endChat();
+                return runtimeContext;
+            }, { omnichannelConfig }),
+        ]);
+
+        const { requestId, conversationDetails } = runtimeContext;
+        const liveWorkItemDetailsRequestUrl = `${omnichannelConfig.orgUrl}/${OmnichannelEndpoints.LiveChatLiveWorkItemDetailsPath}/${omnichannelConfig.orgId}/${omnichannelConfig.widgetId}/${requestId}?channelId=lcw`;
+        const liveWorkItemDetailsResponseDataJson = await liveWorkItemDetailsResponse.json();
+
+        expect(liveWorkItemDetailsRequest.url() === liveWorkItemDetailsRequestUrl).toBe(true);
+        expect(liveWorkItemDetailsResponse.status()).toBe(200);
+        expect(liveWorkItemDetailsResponseDataJson.State).toBe(conversationDetails.state);
+        expect(liveWorkItemDetailsResponseDataJson.ConversationId).toBe(conversationDetails.conversationId);
+        expect(liveWorkItemDetailsResponseDataJson.CanRenderPostChat).toBe(conversationDetails.canRenderPostChat);
+    });
+
+    test('ChatSDK.getConversationDetails() with liveChatContext should not fail', async ({page}) => {
+        await page.goto(testPage);
+
+        const [invalidLiveWorkItemDetailsRequest, invalidLiveWorkItemDetailsResponse, _, liveChatContextLiveWorkItemDetailsRequest, liveChatContextLiveWorkItemDetailsResponse,  runtimeContext] = await Promise.all([
+            page.waitForRequest(request => {
+                return request.url().includes(OmnichannelEndpoints.LiveChatLiveWorkItemDetailsPath);
+            }),
+            page.waitForResponse(response => {
+                return response.url().includes(OmnichannelEndpoints.LiveChatLiveWorkItemDetailsPath);
+            }),
+            await page.evaluate(async ({ omnichannelConfig }) => {
+                const { OmnichannelChatSDK_1: OmnichannelChatSDK } = window;
+                const chatSDK = new OmnichannelChatSDK.default(omnichannelConfig);
+
+                await chatSDK.initialize();
+
+                await chatSDK.startChat();
+
+                const liveChatContext = await chatSDK.getCurrentLiveChatContext();
+
+                const runtimeContext = {
+                    liveChatContext
+                };
+
+                await chatSDK.endChat();
+                runtimeContext.invalidRequestId = chatSDK.requestId;
+
+                const invalidRequestIdConversationDetails = await chatSDK.getConversationDetails();
+                runtimeContext.invalidRequestIdConversationDetails = invalidRequestIdConversationDetails;
+
+                (window as any).runtimeContext = runtimeContext;
+            }, { omnichannelConfig }),
+            page.waitForRequest(request => {
+                return request.url().includes(OmnichannelEndpoints.LiveChatLiveWorkItemDetailsPath);
+            }),
+            page.waitForResponse(response => {
+                return response.url().includes(OmnichannelEndpoints.LiveChatLiveWorkItemDetailsPath);
+            }),
+            await page.evaluate(async ({ omnichannelConfig }) => {
+                const { OmnichannelChatSDK_1: OmnichannelChatSDK, runtimeContext } = window;
+                const chatSDK = new OmnichannelChatSDK.default(omnichannelConfig);
+
+                await chatSDK.initialize();
+
+                const liveChatContextConversationDetails = await chatSDK.getConversationDetails({liveChatContext: runtimeContext.liveChatContext});
+                runtimeContext.liveChatContextConversationDetails = liveChatContextConversationDetails;
+
+                return runtimeContext;
+            }, { omnichannelConfig }),
+        ]);
+
+        const { invalidRequestId, liveChatContext, invalidRequestIdConversationDetails, liveChatContextConversationDetails } = runtimeContext;
+        const invalidLiveWorkItemDetailsRequestUrl = `${omnichannelConfig.orgUrl}/${OmnichannelEndpoints.LiveChatLiveWorkItemDetailsPath}/${omnichannelConfig.orgId}/${omnichannelConfig.widgetId}/${invalidRequestId}?channelId=lcw`;
+        const liveChatContextLiveWorkItemDetailsRequestUrl = `${omnichannelConfig.orgUrl}/${OmnichannelEndpoints.LiveChatLiveWorkItemDetailsPath}/${omnichannelConfig.orgId}/${omnichannelConfig.widgetId}/${liveChatContext.requestId}?channelId=lcw`;
+
+        expect(invalidLiveWorkItemDetailsRequest.url() === invalidLiveWorkItemDetailsRequestUrl).toBe(true);
+        expect(invalidLiveWorkItemDetailsResponse.status()).toBe(400);
+        expect(liveChatContextLiveWorkItemDetailsRequest.url() === liveChatContextLiveWorkItemDetailsRequestUrl).toBe(true);
+        expect(liveChatContextLiveWorkItemDetailsResponse.status()).toBe(200);
+        expect(invalidRequestIdConversationDetails).toEqual({});
+        expect(liveChatContextConversationDetails.state).toBeDefined();
+        expect(liveChatContextConversationDetails.conversationId).toBeDefined();
+        expect(liveChatContextConversationDetails.canRenderPostChat).toBeDefined();
     });
 });