+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Add ability to use `ChatSDK.getLiveChatTranscript()` to fetch live chat transcript from `liveChatContext`
+- Add ability to use `ChatSDK.getConversationDetails()` to fetch conversation details from `liveChatContext`
 
 ### Changed
 - Update `ChatSDKErrors` to include standard ChatSDK errors to be more predictable